Move discord api to separate file
